[change] rename SQL scripts for installing client and group
Use `client.sql` to create the OAuth client in the `auth-server`'s database and
`admin_group.sql` to create the SCIM group `admin` in the `resource-server`'s
database and also associate it with the user with id 1. The old file is still
in place for compatibility reasons, but will receive no further updates and be
eventually removed in a future version. All users are encouraged to update to
the new files.