Stop packaging of test configuration file
The usage of `maven-resources-plugin` in the profile for the ITs, led to
accidental packaging of the `addon-administration.properties` into the
WAR file. This in turn broke the complete app, as the packaged config
file has a higher precedence than the one in `/etc/osiam`.

Solve the issue by copying the config file into the Docker container and
using a Tomcat deployment descriptor to actually load the by Spring.